final tweaks to rendering and refactoring
