+++ demo/test.js
@@ -10,7 +10,6 @@ function updateCharacter() {
 		width: 400,
 		height: 400,
 		showCharacter: false,
-		usePolygonMasks: true,
 	});
 	isCharVisible = true;
 	isOutlineVisible = true;
