+++ src/__tests__/geometry-test.js
@@ -83,24 +83,6 @@ describe('geometry', () => {
     });
   });
 
-  describe('linesToPolygon', () => {
-    it("returns the outline of the polygon that's the stroked path of the points passed in", () => {
-      const points = [
-        new Point(0, 0),
-        new Point(5, 0),
-        new Point(5, 2),
-      ];
-      expect(geometry.linesToPolygon(points, 2)).toEqual([
-        new Point(0, 1),
-        new Point(4, 1),
-        new Point(4, 2),
-        new Point(6, 2),
-        new Point(6, -1),
-        new Point(0, -1),
-      ]);
-    });
-  });
-
   describe('getLineSegmentsPortion', () => {
     it('retuns a new series of points that is portion of the passed in segment', () => {
       // total length 10
